# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /product/{id}:
    get:
      description: Resolves the product URN to a CompanyTwin record and returns the manufacturer Company (twin.manufacturer_company_id).
      operationId: ProductController_getProductCompany
      parameters:
        - name: id
          required: true
          in: path
          description: Product URN (CompanyTwin.asset_ifric_id)
          schema:
            example: urn:product:alpha-machine-001
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get company schema data for a product URN
      tags:
        - Product
  /product/{id}/owner:
    get:
      description: Resolves the product URN to a CompanyTwin record and returns the owner Company (twin.owner_company_id), as opposed to /product/:id which returns the manufacturer.
      operationId: ProductController_getProductOwner
      parameters:
        - name: id
          required: true
          in: path
          description: Product URN (CompanyTwin.asset_ifric_id)
          schema:
            example: urn:product:alpha-machine-001
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get owner company data for a product URN
      tags:
        - Product
  /product/{id}/factory-location:
    get:
      description: Resolves the product URN to a CompanyTwin record, then looks up the Factory referenced by twin.factory_id. Its owner_company_ifric_id matches the company_ifric_id of the company returned by /product/:id/owner.
      operationId: ProductController_getProductFactoryLocation
      parameters:
        - name: id
          required: true
          in: path
          description: Product URN (CompanyTwin.asset_ifric_id)
          schema:
            example: urn:product:alpha-machine-001
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get the factory location for a product URN
      tags:
        - Product
  /product/company-product:
    post:
      operationId: ProductController_addCompanyProduct
      parameters: []
      requestBody:
        required: true
        description: Tags an externally-catalogued product (by its external ID) to a company. Product data lives outside this service — only the ID is stored here.
        content:
          application/json:
            schema:
              type: object
              properties:
                company_ifric_id:
                  type: string
                  example: COMPANY_IFRIC_ID_456
                  description: The unique IFRIC ID of the company
                product_ifric_id:
                  type: string
                  example: urn:product:alpha-machine-001
                  description: External product identifier to tag to the company
                billing_id:
                  type: string
                  example: BILLING_ID_789
                  description: Optional billing ID for the product
              required:
                - company_ifric_id
                - product_ifric_id
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/company/{id}:
    get:
      operationId: ProductController_getCompanyProducts
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/{id}/name:
    get:
      operationId: ProductController_getProductName
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/by-name/{product_name}/id:
    get:
      operationId: ProductController_findProductIdByProductName
      parameters:
        - name: product_name
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/company-product/{id}:
    patch:
      operationId: ProductController_updateCompanyProduct
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Upserts an external product tag for this company, keyed by product_ifric_id.
        content:
          application/json:
            schema:
              type: object
              properties:
                product_ifric_id:
                  type: string
                  example: urn:product:alpha-machine-001
                  description: External product identifier
              required:
                - product_ifric_id
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
    delete:
      operationId: ProductController_deleteCompanyProduct
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin:
    post:
      operationId: ProductController_createCompanyTwin
      parameters: []
      requestBody:
        required: true
        description: Details for creating a company twin
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_company_ifric_id:
                  type: string
                  example: COMPANY_IFRIC_ID_123
                manufacturer_ifric_id:
                  type: string
                  example: MANUFACTURER_IFRIC_ID_456
                asset_ifric_id:
                  type: string
                  example: ASSET_IFRIC_ID_789
                factory_id:
                  type: string
                  example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
                  description: Optional — ties this twin to an existing factory
              required:
                - owner_company_ifric_id
                - manufacturer_ifric_id
                - asset_ifric_id
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
    patch:
      operationId: ProductController_updateCompanyTwin
      parameters: []
      requestBody:
        required: true
        description: Details for updating a company twin
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_company_ifric_id:
                  type: string
                  example: COMPANY_IFRIC_ID_123
                manufacturer_ifric_id:
                  type: string
                  example: MANUFACTURER_IFRIC_ID_456
                asset_ifric_id:
                  type: string
                  example: ASSET_IFRIC_ID_789
                factory_id:
                  type: string
                  example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
                  description: Optional — ties this twin to an existing factory
              required:
                - owner_company_ifric_id
                - manufacturer_ifric_id
                - asset_ifric_id
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin/{id}:
    get:
      operationId: ProductController_getCompanyTwinById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
    delete:
      operationId: ProductController_deleteCompanyTwinAsset
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin/by-asset/{asset_ifric_id}:
    get:
      operationId: ProductController_getCompanyTwinByAssetId
      parameters:
        - name: asset_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin/count:
    post:
      operationId: ProductController_getCompanyTwinCount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin/count/{company_ifric_id}:
    get:
      operationId: ProductController_getCompanyTwinCountByCompanyIfricId
      parameters:
        - name: company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/twin/bulk:
    delete:
      operationId: ProductController_deleteCompanyTwins
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/manufacturer-assets/{id}:
    get:
      operationId: ProductController_getManufacturerAssets
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/manufacturer-owner-assets/{manufacturer_company_ifric_id}/{owner_company_ifric_id}:
    get:
      operationId: ProductController_getManufacturerOwnerAssets
      parameters:
        - name: manufacturer_company_ifric_id
          required: true
          in: path
          schema:
            type: string
        - name: owner_company_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
  /product/owner-assets/{id}:
    get:
      operationId: ProductController_getOwnerAssets
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Product
info:
  title: Ifric Registry Service
  description: Open-source multi-tenant company/user/access-control/digital-twin registry service, split into Auth (session/identity), Company (company CRUD, access groups, physical assets), Product (product catalog, digital twins), Certificate, and Script (one-time seed data) controllers. All use a single bearer-JWT auth mechanism (POST /auth/login, refreshed via POST /auth/refresh). Certificate issuance and company_ifric_id minting integrate with an external open-source ICID-compatible service (see README.md).
  version: 0.1.0
  contact: {}
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
...